Move RAG test to backup in ask yogasutra
